fix(insight): split insight from input [EE-6693]
